+++ deploy/lib/createContainers.js
@@ -180,10 +180,18 @@ module.exports = {
 
     this.serverless.cli.log(`Creating container ${container.name}...`);
 
-    return this.createContainer(params);
+    return this.createContainer(params).then((createdContainer) => {
+      return this.deployContainer(createdContainer.id);
+    });
   },
 
   async updateSingleContainer(container, foundContainer) {
+    // Assign domains to the container before updating it, as it's not possible to manage domains
+    // while the container is updating or pending, and we already wait for the container
+    // to be in a final status before updating it.
+    // => This order of operation is simpler and does not require performing two separate waits.
+    this.applyDomainsContainer(foundContainer.id, container.custom_domains);
+
     let privateNetworkId = container.privateNetworkId;
     const hasToDeletePrivateNetwork =
       foundContainer.private_network_id && !container.privateNetworkId;
@@ -192,7 +200,6 @@ module.exports = {
     }
 
     const params = {
-      redeploy: false,
       environment_variables: container.env,
       secret_environment_variables: await secrets.mergeSecretEnvVars(
         foundContainer.secret_environment_variables,
@@ -225,9 +232,22 @@ module.exports = {
 
     this.serverless.cli.log(`Updating container ${container.name}...`);
 
-    // assign domains
-    this.applyDomainsContainer(foundContainer.id, container.custom_domains);
+    return this.updateContainer(foundContainer.id, params).then(
+      (updatedContainer) => {
+        // If the container is updating, no need to do anything, a redeploy is already in progress.
+        if (
+          updatedContainer.status === "pending" ||
+          updatedContainer.status === "updating"
+        ) {
+          return updatedContainer;
+        }
 
-    return this.updateContainer(foundContainer.id, params);
+        this.serverless.cli.log(
+          `Redeploying container ${container.name} to apply changes...`
+        );
+
+        return this.deployContainer(updatedContainer.id);
+      }
+    );
   },
 };